Fixed Light depth values not being saved to shadowmap.
The lightSpaceView index in the shadow pass was beign set to 2 by the mian renderpass, causing the lightSpaceView used to be unitialized. I created a seperate descriptor just for passing lightSpaceView in the shadow pass and the shadowmap was now filled correctly. Shadows still don't work because currentDepth is always greater than closestDepth in the main scene shader for some reason