Fixed Stop button crash
Moved buffer/image deletion functions out of offscreen deletion queue and created Cleanup functions for Mesh and Light components. Entire offscreen framebuffer no longer needs to be recreated when restarting, only when resizing viewport. There is now an issue where on engine startup the scene is only rendered in part of the viewport with the rest being blank.