Fixed Rendered scene not filling viewport
When the material for each mesh was initialized, its pipeline was set as the initial scene pipeline created with a fixed extent of 1024 by 1024. This caused the scene to render in a square smaller than the viewport. The pipeline of each material then wasn't updated until the entire scene was reloaded.

Now the material pipeline is created with dynamic viewport and scissor which is set for each pipeline in the DrawObjects function. So the pipeline now doesn't have to be recreated each time the viewport is resized, just the framebuffers, attachments and ImTexture descriptor.