Finished Implemeting Line Debug Drawing
Lines can variable length, position and color can now be drawn using the DrawDebugLine function. This will add a pair of vertices to the current frames vertex vector and a draw command to the draw command vector. These are then mapped to their respective buffers and drawn via the vkCmdDrawIndirect command. Next I will work on Implementing Box drawing, probably with Indexed drawing added in to save on having multiple copies of the same vertices.