Added Debug Draw Functions in seperate class
Instead of having to call the draw functions of vulkan engine, the Draw functions form DebugDraw.h can be called instead. These fill global variables with the draw data which are then accessed by the render engine to fill its own data buffers and draw the shapes. This avoids having pointers to the render engine all over the code base.