Implemented Event Manager
Implemented Event Manager for sending messages between systems. For each type of message that you want to send, register it with the ECS. To send an event, call the publish function and to subscribe to the event call the subscribe function with the EventType and pass a shared pointer to the ring buffer you want to store events.