Added AngelScript Scripting Library
AngelScript is open-source scripting library meant to be embedded in applications, such as games. It has C++ style syntax which I prefer and is much easier to integrate with C++ compared to Jinx. I think it is a much better fit for Puffin than Jinx, so I'll likely be removing the latter.