Fixed Crash on Engine Exit, Removed all ComponentEvents, Changed on Component UI is rendered
Fixed the crash when exiting the engine. This was seeming caused by the systems.clear() call in the SystemManager Cleanup function.

Removed all ComponentEvents used for creating/deleting components. I've now gone back to the old way with a BaseComponent class with flags for creating and deleting entities.

Also changed how the Component UI was being rendered, as the Light Component UI wasn't working correctly. Component UI is now all Tree Nodes instead of Collapsing Headers. Also changed some UI colours to make it more interesting.

Script and Physics systems update functions are now called each frame, with a 0.0 passed in as delta time if the game is currently paused or not running. This is to allow initializing and clean up of components when the game is not running.