Moved buffer creation into separate function, added PrepareScene function to ready mesh data for rendering.
Added a new InitBuffers function to handle creation of all buffers for the scene.
Added PrepareScene function. Here the mesh model matrices, vertices, indices and camera data is mapped to their respective buffers ready to be rendered in DeferredRenderer. Currently puffin is hanging when trying to copy the vertex/index data from each meshes buffer to the scene buffers.