Fixed Meshes not being rendered
The scene was not being lit because the result variable in the shader was being filled with garbage data from the light buffers which had no light data in them.
Fixed this by re-adding the light stats buffer which contained the number of active lights for each light type.
Also removed the vector of draw indirect commands from the IndirectBatchDraw struct and replaced it with a count variable for how many indirect commands there are.