Removed bFlagCreated and bFlagDeleted booleans.
Completely removed BaseComponent struct and replaced the two bool's in it with bit set's in the components arrays. Systems now check if components need Initialized or deleted through the ECS world rather than by checking the flags in each component. This should be more cache optimal than the previous Implementation.