Implemented Asset Registry, Asset and StaticMeshAsset
Implemented Asset Registry class to track all assets that have been added to current project. Whenever a class wants to load an asset like a mesh, it will need to get it though the Asset Registry using that assets UUID. MeshComponent has now been setup to store a UUID corresponding to a StaticMeshAsset.