Fixed GetAABB access violation and collision response erros
Removed the virtual GetAABB function declaration from Shape2D for the moment to fix the access violation when calling that function from a derived class. The collision response issue was being caused by copying the transform in the colliders back to the entities, even though this was not being changed during collision detection.