Implemented AssetCache, Updated StaticMeshAsset
The cache will store id/path pairs of all assets registered to the AssetRegistry. When a project is loaded, the asset cache will be used to initialize the registry with all previously used assets to save iterating over the project folder to find all assets at runtime. Also made changes to how vertices/indices are added/retrieved on the StaticMeshAsset