Added PackedArray & PackedVector types. Removed old MarkAsDirty function for entities.
Added new PackedArray & PackedVector types which store a templated type packed together to maximize cache utilization, and which can be indexed with templated id like ECS::Entity.

Removed the old mark to delete functions for Entites and Components and replaced all calls to that code with calls to the Component Flag and new Entity System.