Moved Vertex/Index data out of Mesh Component
Mesh Data is now stored only in the merged vertex/index buffers. Data for rendering each unique mesh is stored in a map with mesh asset UUID's as the key. When there are multiple entities that want to render the same mesh, they are added to the entities set in the mesh render data struct and are then rendered using instanced rendering.